Create password_reset.ps1
This script will reset the user password. If you run the script as an administrator in PowerShell (should have the administrative privileges).

Eg: 
.\password_rest_user.ps1

Enter the username of the account you want to reset the password for: saru
Enter the new password for saru: *******
Confirm the new password for saru: *******
Password for user 'saru' has been successfully reset.